indexer-alt: avoid panic in kv_protocol_configs (#21819)
## Description

Use `ProtocolConfig::get_for_version_if_supported` instead of
`ProtocolConfig::get_for_version`, so that if the version is not
supported, we don't panic and have a chance to propagate an error.

If this pipeline is running in the same process as other pipelines, this
will allow those other pipelines to continue making progress.

## Test plan

New unit test for this behaviour.

```
sui$ cargo nextest run -p sui-indexer-alt
```

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: