[bytecode-verifier] Ready for internal analyzer tool in the nursery (#21573)
## Description 

- When working on the borrow checker, I needed a quick way of analyzing
a corpus of modules and looking at the per-function result. So I wrote a
simple tool to do so
- The actual tool will be checked into the nursery

## Test plan 

- 👀 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: