merge_coins and smash_coins functions on ProgrammableTransactionBuilder (#21893)
## Description 

Added two helper functions `merge_coin` and `merge_all_coins` on
`ProgrammableTransactionBuilder`.

### `merge_coins`

Accepts `destination_coin` and `source_coins` as `ObjectRef` and
`Vec<ObjectRef>`, turns them into argument to perform the `MergeCoin`
command. It returns an `Argument` representing the `destination_coin`.

### `smash_coins`

This is a utility function over `merge_coins` to maximize developer
experience. This function accepts a vector of `ObjectRef` representing
coins, use the first element as `destination_coin` and the next ones as
`source_coins`.

It fails if the provided array is empty. If the array contains only one
element, it skips the `merge_coins` call.

## Test plan 

It has been tested both on local transactions and by adding unit tests.
You can see the tests at the end of the file where the functions have
been implemented.

Note: For an unknown reason some tests crashes in this repo on my
computer, even when running them when being on `main` branch. It seems
to be OS related because I receive errors such as "No such file or
directory". This is my first contribution so I checked at the guidelines
but I didn't found any instruction on how to launch the test. I use the
`cargo test` command and I'm on macOS + Apple Silicon.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [x] Rust SDK: Added new `merge_coins` and `smash_coins` functions to `ProgrammableTransactionBuilder`