Break up big queries that hit limits (#22330)
## Description 

Some transactions are hitting the `[GraphQlError { message: "Query
payload too large: 7424B > 5000B", locations: None, path: None,
extensions: Some(IgnoredAny) }]` error (for whatever the size bigger
than 5000 is), so we are breaking up big multi-get queries.
I am not sure how to compute the size and it's not clear to me what the
query translates to, so I came up with a number that works for the cases
I had and hopefully works in general.
@amnn @stefan-mysten for suggestions if any

## Test plan 

replayed all transaction with that error and they all pass

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: