[static-ptb] Initial check-in of linkage calculation and loading (#22321)
## Description 

Initial check-in of explicit linkage and linkage computation in the
static ptbs.

There is no more `LinkageView` now. Instead there is a `PackageStore`
which allows you to fetch packages based on their object ID, and to
determine type origins, and there is a `LinkedStore` which is just a
`PackageStore` + a `Linkage` -- the `LinkedStore` is what the VM then
uses (it's "where to fetch data, and how to resolve what to fetch from
it"). The creation of a `LinkedStore` is very explicit now, so there is
less ambiguity as to (1) whether you currently have a linkage; and (2)
what that linkage is.

Linkage is computed for calls and publish/upgrade commands and is added
to these commands at the beginning when loading the PTB (as the linkage
is needed for various things like loading the types and the like during
typechecking).

I've also updated the `LoadedFunction` from the VM with the two extra
pieces of metadata to avoid having to re-call back into it (and the
loaded function in the VM already has all this information on it and
we're just throwing it away).

I've de-abstracted a lot of the linkage analysis, however I've left the
core unification and linkage resolution logic from the more extensible
setting, so this should be able to be extended in to e.g., unified
linkage or other linkage analyses that we may want later without too
much pain.

There is currently a lot of "VM type <-> Adapter type <-> TypeTag <->
MoveObjectType" type of conversions in a number of places. For now we
will just deal with these conversions but it would be good to go through
this once we concretize more things to make sure we are not converting
places where we don't need to be.


## Test plan 

Nothing yet -- that will come once we get everything hooked up :) 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: