[move][easy] Filter out macro-locations when computing source coverage (#22409)
## Description 

Expanded macros pull in source locations from other files. This can lead
to errors when computing source coverage (spans that refer to locations
outside of the current file).

When computing the coverage map for a file, we remove any locations not
located within that file (i.e., locations from macro expansion). Sadly
this means that we lose some level of coverage information about the
macro call, but hopefully in the future with some nicer locations we can
improve this. But for now, this at least stops it from erroring out
completely.

## Test plan 

CI + tested it on a local repro.