Save packages for tracing when publishing or upgrading (#22511)
## Description 

Save locally published and upgraded packages so tracing of init has code
to refer to

## Test plan 

Test locally as we do not have a way to test it yet in replay 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: