[consensus] flush all pending commits (#22628)
## Description 

It is unnecessary to restrict commit flushing to finalized commits only.
Sui / execution replays consensus from last processed commit, and
replaying commits rerun the transaction finalization logic. Also it is
incorrect to flush blocks without flushing the commit that sets the GC
round at the time. The GC round determines which ancestors of a block
must exist, so it should be flushed together with the ancestors before
persisting a block.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: