[sui CLI] Remove old `profile-transaction` and `replay-checkpoint` cmds (#22644)
## Description 

These commands were not usable anymore, and we plan on supporting the
capabilities of `profile-transaction` in a different way, and
`replay-checkpoint` is going to go away entirely.

## Test plan 

CI + 👀 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [X] CLI: Remove `profile-transaction` and `replay-checkpoint` commands
from the `sui client` commands.
- [ ] Rust SDK: