allow test_scenario to set context values (#22607)
## Description 

Make a builder style API on test scenario.
Implementation is not really like a builder as the `TxContext` is
changed in-place. That is, however, the simpler and imo cleaner way to
do it albeit rather verbose.
I was defining the API a bit more "rigorous" (as in allowing certain
changes like rgp only on epoch changes) but then decided the complexity
was not worth it and it's a test API that could be made more permissive.
`TxContext` exposes few more public test only API. `ids_created` is now
a public test only function and there is also a `get_ids_cretated`. That
seems all a bit too much and not sure what I should do. I can keep the
`get_ids_cretated` and remove `ids_cretated` and call it a day, I can
deprecate `get_ids_cretated` which is public and cannot be removed, or
keep it both. Suggestions welcome.
This PR removes the changes made for the added API in #22341
(`gas_price` and `reference_gas_price`) in `TxContext::new` and makes
things a bit cleaner. They should be on the same protocol version so we
should be good

## Test plan 

Added test to check the values in `TxContext` are set and changed

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: