[move][tracing] add storeage ID to open frame event (#22670)
## Description 

What it says on the title

## Test plan 

Ran a replay locally and checked the output is what I expected

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------

Co-authored-by: Adam Welc <adam@mystenlabs.com>
Co-authored-by: Dario Russi <113150618+dariorussi@users.noreply.github.com>