[traffic controller] downgrade shutdown error case to warn! log (#22782)
## Description 

> looks like it got triggered because tallies weren't able to process
during what seems to be a slow process shutdown. Since this is the first
time this triggers I'm inclined to leave it as is, but if we notice it
happens more in these shutdown situations, I will downgrade this to a
warn! log

Not an actionable alert as is. 



---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: