[trace-view] Fixed a problem related to using incorrect pkg IDs (#22800)
## Description 

This PR fixes a problem with replay and trace-debugging transactions
that include calls into upgraded packages. The problem was related to
modules being represented as belonging to the original package rather
than the defining (upgraded) package. This was manifested in two places:
- `OpenFrame` event in the trace
- debug info

The result of the former was that a function that did not exist in the
original package could not be found during trace debugging as the
debugger, seeing module name and original package ID, would look for
this function in debug info representing a module from the original
package, where this function indeed does not exist.

The result of the latter was that debug info (key-ed in the debugger on
the module name and package ID) for different versions of the module
could not be distinguished in the debugger as all these debug info-s
would contain original package ID

## Test plan 

Tested that upgraded packages are handled correctly, In particular,
mainnet transaction with the following digest can now be trace-debugged
correctly: `8jfAyyiQZG895NtjBjnb3eEjiwQtSwXi3FYHMGuJJZr6