[sui-framework-snapshot] Set full shas where possible (#22829)
## Description 

This PR updates the short shas to full shas where possible in the
already existing `sui-framework-snapshot/manifest.json`. The reason for
this change is that the new pkg management system benefits from using
full shas to avoid downloading the whole history to be able to checkout
with only the short sha.

It also adds a new `GIT_FULL_SHA_REVISION` macro on `bin-version` crate
to allow us to get the full 40 characters SHA to be used in
`update_snapshot` in the `sui-framework-snapshot` crate.

## Test plan 

Existing tests

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: