[static-ptbs] linkage fixes (#22861)
## Description 

Bugfixes for the static PTBs

1. Bottom commit adds a missing trait implementation. Before in the
static PTBs this was falling back to a default trait implementation
which didn't do the correct thing.
2. Top commit changes the way we load types. Before we would compute a
unified linkage for the type and then load that, however this poses
linking issues. Instead we use the way we load types today where we load
from the leaves up.

## Test plan 

CI + replay + new test