consistency: associate snapshots with watermarks
## Description

Part of the work to integrate the consistent store with GraphQL. In
particular, GraphQL needs to know the high watermark (checkpoint, epoch,
transaction, timestamp) for the consistent service. The `AvailableRange`
method only returns the min and max checkpoints.

This change associates a full watermark with each snapshot so that (in a
later change) we can augment `AvailableRange`.

Part of this change also fixes a bug where synchronizers could start out
of sync, because a different `next_snapshot_checkpoint` was passed to
each one (if subsequent pipelines have increasing high watermarks).

## Test plan

```
$ cargo nextest run -p sui-indexer-alt-consistent-store
```