enforce 0600 flags on sui.keystore (#23036)
## Description 

Ensure that (like SSH) we reduce the file permissions on sui.keystore.

Note that this will automatically alter the flags of the sui.keystore
file upon load or save.

This change does not apply to Windows.

## Test plan 

CI Pipeline and manual testing.

---

## Release notes

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes):
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [x] CLI: (linux, macos) Ensure that sui.keystore files are flagged
with 0600 flags. This is a defense-in-depth measure to defend Sui
wallets running on shared multi-user hosts.
- [ ] Rust SDK: